no hope
