+++ FPTManagerSutdent/Views/Students/Create.cshtml
@@ -25,6 +25,11 @@
                 <span asp-validation-for="Email" class="text-danger"></span>
             </div>
             <div class="form-group">
+                <label asp-for="Password" class="control-label"></label>
+                <input asp-for="Password" class="form-control" />
+                <span asp-validation-for="Password" class="text-danger"></span>
+            </div>
+            <div class="form-group">
                 <label asp-for="Gender" class="control-label"></label>
                 <select asp-for="Gender" class="form-control">
                     @foreach (var gender in Enum.GetValues(typeof(GenderStudent)))
@@ -46,7 +51,7 @@
             </div>
             <div class="form-group">
                 <label asp-for="DoB" class="control-label"></label>
-                <input asp-for="DoB" class="form-control" type="date" />
+                <input asp-for="DoB" class="form-control" />
                 <span asp-validation-for="DoB" class="text-danger"></span>
             </div>
             <div class="form-group">
@@ -69,6 +74,3 @@
     <a asp-action="Index">Back to List</a>
 </div>
 
-@section Scripts {
-    @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
-}
