+++ FPTManagerSutdent/Views/Students1/Delete.cshtml
@@ -1,4 +1,4 @@
-﻿@model FPTManagerSutdent.Models.Account
+﻿@model FPTManagerSutdent.Models.Student
 
 @{
     ViewData["Title"] = "Delete";
@@ -8,9 +8,15 @@
 
 <h3>Are you sure you want to delete this?</h3>
 <div>
-    <h4>Account</h4>
+    <h4>Student</h4>
     <hr />
     <dl class="dl-horizontal">
+        <dt>
+            @Html.DisplayNameFor(model => model.Name)
+        </dt>
+        <dd>
+            @Html.DisplayFor(model => model.Name)
+        </dd>
         <dt>
             @Html.DisplayNameFor(model => model.Email)
         </dt>
@@ -28,6 +34,30 @@
         </dt>
         <dd>
             @Html.DisplayFor(model => model.Salt)
+        </dd>
+        <dt>
+            @Html.DisplayNameFor(model => model.Gender)
+        </dt>
+        <dd>
+            @Html.DisplayFor(model => model.Gender)
+        </dd>
+        <dt>
+            @Html.DisplayNameFor(model => model.Phone)
+        </dt>
+        <dd>
+            @Html.DisplayFor(model => model.Phone)
+        </dd>
+        <dt>
+            @Html.DisplayNameFor(model => model.Address)
+        </dt>
+        <dd>
+            @Html.DisplayFor(model => model.Address)
+        </dd>
+        <dt>
+            @Html.DisplayNameFor(model => model.DoB)
+        </dt>
+        <dd>
+            @Html.DisplayFor(model => model.DoB)
         </dd>
         <dt>
             @Html.DisplayNameFor(model => model.CreatedAt)
